fixed tuple reference problem.
as the tuple is a temporary, it can't hold lvalue references, only rvalue references. Therefore, a and b (inside the tuple) become rvalue references as well. This still allows the user to modify the original data through this rvalue reference object.

See the following loop:

for (auto [a, b] : view2)
    {

    }

a and b are rvalue references